Creacion e Inicialización del Componente y Estados
