--- src/pages/Comentarios/COMENT.css
@@ -1,126 +1,0 @@
-h1 {
-    color: #031026;
-    font-size: 2.2rem;
-    margin-bottom: 1rem;
-    font-weight: 700;
-  }
-  
-  h2 {
-    color: #468189;
-    font-size: 2rem;
-    margin-bottom: 0.8rem;
-    font-weight: 600;
-  }
-  
-  h3 {
-    color: #77ACA2;
-    font-size: 1.3rem;
-    margin-bottom: 0.5rem;
-    font-weight: 500;
-  }
-
-.coment-container {
-    max-width: 800px;
-    margin: 2rem auto;
-    padding: 1rem;
-    background-color: #F4E9CD; /* Fondo general */
-    font-family: 'Segoe UI', sans-serif;
-    border-radius: 12px;
-    box-shadow: 0 2px 10px rgba(3, 16, 38, 0.1);
-    
-    /* Centrar el contenido */
-    display: flex;
-    flex-direction: column; /* Asegura que los elementos se apilen verticalmente */
-    justify-content: center; /* Centra verticalmente */
-    align-items: center; /* Centra horizontalmente */
-    height: 100vh; /* Asegura que ocupe toda la altura de la ventana */
-}
-
-.coment-textarea {
-    width: 50%;
-    min-height: 100px;
-    padding: 12px;
-    border: 2px solid #468189;
-    border-radius: 8px;
-    font-size: 1rem;
-    resize: vertical;
-    background-color: #fffefb;
-    color: #031026;
-    margin-bottom: 1rem;
-    transition: border 0.3s;
-}
-
-.coment-button {
-    background-color: #468189;
-    color: #fff;
-    border: none;
-    padding: 10px 28px;
-    font-size: 1rem;
-    border-radius: 8px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-    margin-top: 1rem; /* Espaciado entre el área de texto y el botón */
-}
-
-.coment-button:hover {
-    background-color: #36636b;
-}
-
-.coment-list {
-    margin-top: 2rem;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center; /* Centra los elementos de la lista */
-}
-
-.coment-item {
-    display: flex;
-    padding: 1rem;
-    background: #9DBEBB;
-    border-radius: 10px;
-    box-shadow: 0 2px 6px rgba(3, 16, 38, 0.1);
-    margin-bottom: 1.5rem;
-    width: 80%; /* Ajusta el ancho de los comentarios */
-    transition: transform 0.2s;
-}
-
-.coment-item:hover {
-    transform: translateY(-2px);
-}
-
-.coment-image-preview {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 10px;
-  }
-  
-  .coment-image-box {
-    width: 120px; /* Tamaño fijo para la caja de la imagen */
-    height: 120px; /* Tamaño fijo para la caja de la imagen */
-    overflow: hidden;
-    border: 1px solid #ccc;
-    position: relative;
-  }
-  
-  .coment-image {
-    width: 100%; /* Ajusta la imagen al tamaño del contenedor */
-    height: 100%;
-    object-fit: cover; /* Mantiene la proporción de la imagen y la ajusta al contenedor sin distorsionarla */
-  }
-  
-  .remove-image-btn {
-    position: absolute;
-    top: 5px;
-    right: 5px;
-    background: red;
-    color: white;
-    border: none;
-    padding: 5px;
-    cursor: pointer;
-  }
-  
-  .remove-image-btn:hover {
-    background: darkred;
-  }
-
